+++ docs/source/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+ - new model MergeIndexImageModel
+
+
 ## 2.1.0 - 2021-01-20
 
 ### Added